+++ types/haltian/thingseePresence/uplink.spec.js
@@ -1,0 +1,70 @@
+const chai = require("chai");
+const { validate } = require("jsonschema");
+const rewire = require("rewire");
+const utils = require("test-utils");
+
+const { assert } = chai;
+
+describe("Haltian Thingsee Presence Sensor Uplink", () => {
+  let moveCountSchema = null;
+  let consume = null;
+  before((done) => {
+    const script = rewire("./uplink.js");
+    consume = utils.init(script);
+    utils
+      .loadSchema(`${__dirname}/move_count.schema.json`)
+      .then((parsedSchema) => {
+        moveCountSchema = parsedSchema;
+        done();
+      });
+  });
+
+  let lifecycleSchema = null;
+  before((done) => {
+    utils
+      .loadSchema(`${__dirname}/lifecycle.schema.json`)
+      .then((parsedSchema) => {
+        lifecycleSchema = parsedSchema;
+        done();
+      });
+  });
+
+  describe("consume()", () => {
+    it("should decode the Haltian Thingsee Presence Sensor occupancy payload", () => {
+      const data = {
+        data: {
+          tsmId: 2100,
+          tsmTuid: "TSPR04EZU20400472",
+          tsmTs: 1657195118,
+          tsmGw: "TSGW06EWK14800496",
+          tsmEv: 10,
+          deploymentGroupId: "prch00switzerland",
+          state: 0,
+        },
+      };
+      utils.expectEmits((type, value) => {
+        assert.equal(type, "sample");
+        assert.isNotNull(value);
+        assert.typeOf(value.data, "object");
+
+        assert.equal(value.topic, "occupancy");
+        assert.equal(value.data.occupancy, 0);
+
+        validate(value.data, moveCountSchema, { throwError: false });
+      });
+
+      utils.expectEmits((type, value) => {
+        assert.equal(type, "sample");
+        assert.isNotNull(value);
+        assert.typeOf(value.data, "object");
+
+        assert.equal(value.topic, "lifecycle");
+        assert.equal(value.data.reason, "TIME");
+
+        validate(value.data, lifecycleSchema, { throwError: true });
+      });
+
+      consume(data);
+    });
+  });
+});